backend/pyproject.toml
[project]
name = "store-api"
version = "0.1.0"
description = ""
authors = [
    {name = "yeelisey"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "minio (>=7.2.17,<8.0.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
